Covariance Matrix Calculation for (Potentially Grouped) Data

Covariance Matrix Calculation for (Potentially Grouped) Data

cov(x, ..., covEst = stats::cov)

Arguments

x
data you would like the covariance matrix of. The parameter x can be a data.frame, grouped_df, resample, or matrix.

...
other options passed to covarince estimation method

covEst
covariance or precision matrix estimation method, as a function. Defaults to stats::cov.

Value

list of matries with class covariance by group. These matrices may be covariance or precision matrices, depending on the function supplied to the covEst argument.

Details

This function has the capability to calculate total and group-level covariance and precision matrices estimated from the given data. Specify group membership with the group = "COLNAME" syntax. Additionally, this function returns objects with an additional class: when this function returns group-level lists of matrices, each covariance or precision matrix will have the R classes matrix and covariance; likewise, total covariance or precision matrices are also returned with the classes matrix and covariance. Moreover, these matrices have an additional named attribute: df, for degrees of freedom. The computational functionality of the covariance class will be explored in future updates.

Examples

cov(iris[,1:4])

#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    0.6856935  -0.0424340    1.2743154   0.5162707
#> Sepal.Width    -0.0424340   0.1899794   -0.3296564  -0.1216394
#> Petal.Length    1.2743154  -0.3296564    3.1162779   1.2956094
#> Petal.Width     0.5162707  -0.1216394    1.2956094   0.5810063
#> attr(,"df")
#> ~149
#> <environment: namespace:covEstR>
#> attr(,"class")
#> [1] "covariance" "matrix"    

cov(iris, group = "Species")

#> $setosa
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.12424898 0.099216327  0.016355102 0.010330612
#> Sepal.Width    0.09921633 0.143689796  0.011697959 0.009297959
#> Petal.Length   0.01635510 0.011697959  0.030159184 0.006069388
#> Petal.Width    0.01033061 0.009297959  0.006069388 0.011106122
#> attr(,"df")
#> ~49
#> <environment: 0x10caf14d0>
#> attr(,"class")
#> [1] "covariance" "matrix"    
#> 
#> $versicolor
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.26643265  0.08518367   0.18289796  0.05577959
#> Sepal.Width    0.08518367  0.09846939   0.08265306  0.04120408
#> Petal.Length   0.18289796  0.08265306   0.22081633  0.07310204
#> Petal.Width    0.05577959  0.04120408   0.07310204  0.03910612
#> attr(,"df")
#> ~49
#> <environment: 0x10caf14d0>
#> attr(,"class")
#> [1] "covariance" "matrix"    
#> 
#> $virginica
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.40434286  0.09376327   0.30328980  0.04909388
#> Sepal.Width    0.09376327  0.10400408   0.07137959  0.04762857
#> Petal.Length   0.30328980  0.07137959   0.30458776  0.04882449
#> Petal.Width    0.04909388  0.04762857   0.04882449  0.07543265
#> attr(,"df")
#> ~49
#> <environment: 0x10caf14d0>
#> attr(,"class")
#> [1] "covariance" "matrix"    
#> 

cov(iris[,1:4], covEst = Haff_shrinkage)

#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    1462.2325   -950.2355   -1035.3471    812.4760
#> Sepal.Width     -950.2355   1567.5041     917.3115   -873.4804
#> Petal.Length   -1035.3471    917.3115    1422.1717  -2054.3881
#> Petal.Width      812.4760   -873.4804   -2054.3881   3922.1784
#> attr(,"df")
#> ~149
#> <environment: namespace:covEstR>
#> attr(,"class")
#> [1] "covariance" "matrix"